--- client/src/AdminLogin.jsx
@@ -1,50 +1,0 @@
-import React, { useState } from "react";
-
-const AdminLogin = ({ onLogin }) => {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-
-  const handleLogin = async () => {
-    try {
-      // Pass both username and password in the login request
-      await onLogin({ username, password });
-      // Clear username and password after successful login
-      setUsername("");
-      setPassword("");
-      setError("");
-    } catch (error) {
-      console.error("Login failed:", error);
-      setError("Login failed. Please try again.");
-    }
-  };
-
-  return (
-    <div>
-      <h2>Login</h2>
-      <label>
-        Username:
-        <input
-          type="text"
-          value={username}
-          onChange={(e) => setUsername(e.target.value)}
-        />
-      </label>
-      <br />
-      <label>
-        Password:
-        <input
-          type="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-        />
-      </label>
-      <br />
-      <button onClick={handleLogin}>Login</button>
-      {/* Text is red when an error occurs */}
-      {error && <p style={{ color: "red" }}>{error}</p>}
-    </div>
-  );
-};
-
-export default AdminLogin;
